1.0.0 pre2.3 (#19)
*breaking changes*

* `getMenuOptions` has been renamed to `getMenuItems` in all places it occurs

Additional Changes:

* Automated tests setup with travis + saucelabs
* Removes SystemBus from `root.js`; uses `Layer.Utils.defer` instead
* Memory leaks removed from unit tests
* Upgrade to jasmine 3.0.0
* Refactored tests and CI with Saucelabs
* WEB-1680: `messageStatusRenderer` and `dateRenderer` properties now used on initial load as well as new messages
* WEB-1685: Improves quality and size of Preview Images, and adds static properties to let developers customize preview sizing and quality
* WEB-1648: Replace `MessageTypeModel.responses` object with `MessageTypeModel.responses` a `MessageTypeResponseSummaryModel` instance (i.e. added the `MessageTypeResponseSummaryModel` class)
    * Adds `model.responses.getResponse(responseName, identityId)`
    * Adds `model.responses.getResponses(responseName, identityIds)`
* Redefines the `layer-widget-destroyed` event; it now triggers on `document.body` for each removed component; access `evt.detail.target` not `evt.target` to determine what UI Component has been removed and is about to be destroyed.
* Fixes build script that strips HTML Comments out of templates
* ReplaceableContent subproperties now accept `null` values as a way to prevent anything from being rendered in an area.
* Adds the 'layer-groups.css' theme
* Flexbox workarounds added to the CSS
* Now supports setting `model.action.event = null;` to prevent a model's action from triggering

+++ src/ui/mixins/tests/has-query-test.js
@@ -38,7 +38,7 @@ describe("Has Query Mixin", function() {
     });
     query.isFiring = false;
     query.data = [client.user];
-    for (i = 0; i < 100; i++) {
+    for (i = 0; i < 25; i++) {
       query.data.push(
         new Layer.Core.Identity({
               userId: 'user' + i,
@@ -56,7 +56,8 @@ describe("Has Query Mixin", function() {
 
   afterEach(function() {
     jasmine.clock().uninstall();
-    client.destroy();
+    if (el) el.destroy();
+    if (client) client.destroy();
     document.body.removeChild(testRoot);
 
   });
